% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.area.R
\name{core.area}
\alias{core.area}
\title{Derive the core area isopleth from a utilization distribution using the method described in Vander Wal & Rogers (2012).}
\usage{
core.area(r, plot = FALSE, silent = TRUE)
}
\arguments{
\item{r}{spatRaster (utilization distribution)}

\item{plot}{logical indicating whether or not to print result}

\item{silent}{logical indicating whether or not to print messages (default = TRUE)}
}
\value{
isopleth associated with core/intensive use area of input UD
}
\description{
Derive the core area isopleth from a utilization distribution using the method described in Vander Wal & Rogers (2012).
}
