% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1score.R
\name{F1score}
\alias{F1score}
\title{F1 Score for binned continuous data (kernel utilization distributions)}
\usage{
F1score(pred, obs, threshVal = 50, nbins = 10, beta = 1)
}
\arguments{
\item{pred}{spatRaster object corresponding to predicted probabilities of occurrence (wmKDE()$iso) from fitted wmKDE (training set) model (0 <= x >= 100).}

\item{obs}{spatRaster object corresponding to observed probabilities of occurrence (wmKDE()$iso) from fitted wmKDE (validation set) model (0 <= x >= 100).}

\item{threshVal}{minimum probability value considered as a 'presence' (TRUE).}

\item{nbins}{number of bins to create for the confusion matrix (default = 10).}

\item{beta}{relative weight of recall relative to precision. Default is equal weight to both.}
}
\value{
list containing 1) recall, 2) precision, and 3) F1 score for the evaluated data (predicted vs observed) and input parameters.
}
\description{
F1 Score for binned continuous data (kernel utilization distributions)
}
