% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhat2confin.R
\name{fhat2confin}
\alias{fhat2confin}
\title{Convert UD density estimates into relative probabilities}
\usage{
fhat2confin(z)
}
\arguments{
\item{z}{numeric vector representing UD density values (fhat)}
}
\value{
numeric vector of length(z) representing isopleths corresponding to input density values
}
\description{
Convert UD density estimates into relative probabilities
}
