% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{finetune}
\alias{finetune}
\title{Redistribute UD probability densities proportionally so as to ensure that the sum of cell volumes equals one.}
\usage{
finetune(r, out = 1)
}
\arguments{
\item{r}{spatRaster containing kernel density probabilities (utilization distribution)}

\item{out}{integer parameter specifying whether output cell values are aspatial (out=1) or volume-based (out=2).}
}
\value{
spatRaster with probabilities adjusted such that corresponding volumes sum to exactly 1
}
\description{
Redistribute UD probability densities proportionally so as to ensure that the sum of cell volumes equals one.
}
