% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hone.extent.R
\name{hone.extent}
\alias{hone.extent}
\title{hone.extent
Generate a buffered Extent around a polygon or other spatial object from which an extent object can be extracted.}
\usage{
hone.extent(spatpol, r = 1000)
}
\arguments{
\item{spatpol}{Spatial polygon from which to generate the extent object}

\item{r}{expansion radius in projected coordinates (m)}
}
\value{

}
\description{
hone.extent
Generate a buffered Extent around a polygon or other spatial object from which an extent object can be extracted.
}
