% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isopolygonize.R
\name{isopolygonize}
\alias{isopolygonize}
\title{Given a Utilization Distribution, generate polygon(s) corresponding to specified probability contours}
\usage{
isopolygonize(r, sproj = NULL, prob = 0.95, spType = "poly")
}
\arguments{
\item{r}{spatRaster object representing a Utilization Distribution.}

\item{sproj}{projected coordinate system of input UD}

\item{prob}{probability contour(s) at which to generate the polygon}

\item{spType}{target output geometry type (sf object). Default is 'poly', otherwise 'line'.}
}
\value{
sf polygons object representing isopleth contours of input UD
}
\description{
Given a Utilization Distribution, generate polygon(s) corresponding to specified probability contours
}
