% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.grid.R
\name{kernelGrid}
\alias{kernelGrid}
\title{Define a spatial grid corresponding to the distribution area of the population of interest.}
\usage{
kernelGrid(locs, exp.range = 3, cell.size)
}
\arguments{
\item{locs}{data.frame with columns 'x' and 'y' corresponding to projected spatial coordinates of relocations}

\item{exp.range}{single integer defining the range expansion parameter ensuring sufficient area is available to fully estimate the population UD (3 by default).}

\item{cell.size}{desired cell resolution (dx=dy)}
}
\value{
Generates range.x and grid.size arguments required by KernSmooth::bkde2D.
}
\description{
Function called by mwKDE to define the common grid over which individual UDs are estimated.
}
