% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwKDE.R
\name{mwKDE}
\alias{mwKDE}
\title{Mean weighted kernel density estimation}
\usage{
mwKDE(
  xy,
  avg = TRUE,
  id,
  spwt = TRUE,
  wts = NULL,
  bw.global = TRUE,
  zscale = TRUE,
  spatres = 1000,
  ncores = ifelse(avg, parallel::detectCores() - 1, 1),
  ow = TRUE,
  titre = NULL,
  export = TRUE,
  obj.ret = FALSE,
  herd.grid = NULL
)
}
\arguments{
\item{xy}{2 column matrix or data.frame corresponding to projected GPS relocation coordinates (x, y).}

\item{avg}{logical indicating whether an averaged UD is desired. When avg=FALSE, only one kernel is generated using all relocations.}

\item{id}{vector of length equal to nrow(xy) specifying the unique collared animal from which each relocated was transmitted. Only necessary when avg=TRUE.}

\item{spwt}{logical indicating whether weights should be applied to individual relocations.}

\item{wts}{vector equal in length to nrow(xy) that defines the spatial weights applied to individual relocations. If the sum of relocation weights does not equal the sample size, weights are adjusted accordingly and a warning is issued. Only applicable when spwt=TRUE.}

\item{bw.global}{logical indicating whether bandwidth smoothing should be derived from all relocations (recommended) or made to vary according to individual sample (point pattern) distributions.}

\item{zscale}{logical indicating whether individual UD density values should be rescaled prior to cellwise averaging. Only applicable when avg=TRUE.}

\item{spatres}{vector specifying the desired spatial resolution of the output UD in the x & y dimensions. Arguments of length 1 are recycled.}

\item{ncores}{integer indicating the number of parallel processes (threads) over which to execute the estimation of individual UDs. Defaults to detectCores()-1 when avg=TRUE, otherwise 1.}

\item{ow}{logical; overwrite existing files?}

\item{titre}{optional title to print to plot}

\item{export}{logical; write results to file?}

\item{obj.ret}{logical; should results be returned?}

\item{herd.grid}{optional specification of grid over which to estimate the UD. Generated with kernel.grid().}
}
\value{

}
\description{
Estimate the mean utilization distribution (UD) of a wildlife population given GPS collar telemetry relocations derived from individually monitored sample animals.
Two levels of weighting are possible:
\enumerate{
\item Individual GPS relocations may be weighted during kernel density estimation to account for variation in data quality and/or habitat transformation over time.
\item Individual UDs may be weighted during cell-wise averaging to account for individual variation in sample size and/or quality.
}
}
