% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwUD.R
\name{mwUD}
\alias{mwUD}
\title{Cell-wise weighted means of multiple Utilization Distributions}
\usage{
mwUD(UDlist, w, dres = NULL, silent = FALSE, dcrs = NULL, checksum = TRUE)
}
\arguments{
\item{UDlist}{list of utilization distributions of equal dimensions.}

\item{w}{vector of individual UD weights for cell-wise weighted means.}

\item{dres}{desired output spatial resolution of mean weighted UD in projected coordinate units (m)s.}

\item{silent}{logical; should messages be printed?}

\item{dcrs}{projected coordinate system of input UDs}

\item{checksum}{logical; should adjustments be made to ensure sure that all cell volumes sum to 1?}
}
\value{

}
\description{
Cell-wise weighted means of multiple Utilization Distributions
}
