% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.pts.in.poly.R
\name{percent.pts.in.poly}
\alias{percent.pts.in.poly}
\title{Calculate the percentage of relocations situated within a given polygon.}
\usage{
percent.pts.in.poly(xy, poly, w = NULL)
}
\arguments{
\item{xy}{2-column data.frame or matrix containing projected x and y coordinates}

\item{poly}{a SpatialPolygons or SpatialPolygonsDataFrame object}

\item{w}{optional vector of weights equal in length to nrow(xy)}
}
\value{

}
\description{
Calculate the percentage of relocations situated within a given polygon.
}
