% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgoi.R
\name{pgoi}
\alias{pgoi}
\title{Compute the probabilistic general overlap index as per McIntyre & Barry (2018). “A Lattice-Based Smoother for Regions With Irregular Boundaries and Holes.” Journal of Computational and Graphical Statistics 27, no. 2 (2018): 360–67. https://doi.org/10.1080/10618600.2017.1375935.}
\usage{
pgoi(r)
}
\arguments{
\item{r}{A spatRaster object consisting of exactly 2 layers, the values of which are to be compared}
}
\value{
numeric pgoi
}
\description{
Compute the probabilistic general overlap index as per McIntyre & Barry (2018). “A Lattice-Based Smoother for Regions With Irregular Boundaries and Holes.” Journal of Computational and Graphical Statistics 27, no. 2 (2018): 360–67. https://doi.org/10.1080/10618600.2017.1375935.
}
