% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast2UD.R
\name{rast2UD}
\alias{rast2UD}
\title{Convert a RasterLayer object to a UD object}
\usage{
rast2UD(r)
}
\arguments{
\item{r}{RasterLayer object}
}
\value{
UD object retaining cell values of original RasterLayer object at original grid cell locations.
}
\description{
Convert a RasterLayer object to a UD object
}
