% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweightedUD.R
\name{reweightedUD}
\alias{reweightedUD}
\title{Redistribute UD probability densities proportionally so as to ensure that the sum of cell volumes equals one.}
\usage{
reweightedUD(UD, out = 1)
}
\arguments{
\item{UD}{a list containing 3 elements: 'x1' (x coordinates), 'x2' (y coordinates), 'fhat' (UD probability densities).}

\item{out}{integer parameter specifying whether output cell values are aspatial (out=1) or volume-based (out=2).}
}
\description{
Redistribute UD probability densities proportionally so as to ensure that the sum of cell volumes equals one.
}
