% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweightUD.R
\name{reweightUD}
\alias{reweightUD}
\title{Redistribute UD probability densities proportionally so as to ensure that the sum of cell volumes equals one.}
\usage{
reweightUD(UD, out = 1)
}
\arguments{
\item{UD}{a list containing 3 elements: 'x1' (x coordinates), 'x2' (y coordinates), 'fhat' (UD probability densities).}

\item{out}{integer parameter specifying whether output cell values are aspatial (out=1) or volume-based (out=2).}
}
\value{
original input UD with probabilities adjusted such that corresponding volumes sum to exactly 1
}
\description{
Redistribute UD probability densities proportionally so as to ensure that the sum of cell volumes equals one.
}
