% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetDB.R
\name{subsetDB}
\alias{subsetDB}
\title{Subset a GPS collar telemetry database according to user preferences and/or analytical objectives.}
\usage{
subsetDB(
  df,
  pop = NULL,
  saison = NULL,
  sex = NULL,
  annee = NULL,
  id = NULL,
  nmin = 50
)
}
\arguments{
\item{df}{data.frame}

\item{pop}{factor}

\item{saison}{factor}

\item{sex}{factor}

\item{annee}{integer}

\item{id}{character}

\item{nmin}{vector}
}
\value{
data.frame
}
\description{
Subset a GPS collar telemetry database according to user preferences and/or analytical objectives.
}
