% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD2rast.R
\name{ud2rast}
\alias{ud2rast}
\title{Convert a UD object to a RasterLayer object}
\usage{
ud2rast(UD, sproj = NULL)
}
\arguments{
\item{UD}{input utilization distribution (UD) object; a list containing three elements named 'x1', 'x2' and 'fhat'.}

\item{sproj}{projected coordinate system of the input UD.}
}
\value{
spatRaster object retaining original UD cell values at original xy grid cell locations.
}
\description{
Convert a UD object to a RasterLayer object
}
