% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD2sf.R
\name{UD2sf}
\alias{UD2sf}
\title{Given a Utilization Distribution, generate polygon(s) corresponding to specified probability contours}
\usage{
UD2sf(UD, sproj = NULL, probs = 0.95, spType = "poly")
}
\arguments{
\item{UD}{object representing a Utilization Distribution. A list containing 3 elements: x (x1) and y (x2) coordinates, and corresponding density estimates (fhat)}

\item{sproj}{projected coordinate system of input UD}

\item{probs}{probability contour(s) at which to generate the polygon}

\item{spType}{target output geometry type (sf object). Default is 'poly', otherwise 'line'.}
}
\value{
sf polygons object representing isopleth contours of input UD
}
\description{
Given a Utilization Distribution, generate polygon(s) corresponding to specified probability contours
}
