% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD2sp.R
\name{UD2sp}
\alias{UD2sp}
\title{Given a Utilization Distribution, generate polygon(s) corresponding to specified probability contours}
\usage{
UD2sp(UD, sproj = NULL, probs = 0.95)
}
\arguments{
\item{UD}{object representing a Utilization Distribution. A list containing 3 elements: x (x1) and y (x2) coordinates, and corresponding density estimates (fhat)}

\item{sproj}{projected coordinate system of input UD}

\item{probs}{probability contour(s) at which to generate the polygon}
}
\description{
Given a Utilization Distribution, generate polygon(s) corresponding to specified probability contours
}
