% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Compute a suite of common metrics serving to assess predictive capacity of an empirical/training model given an observed/validation sample}
\usage{
validate(r, probs = c(50, 70))
}
\arguments{
\item{r}{a spatRaster consisting of exactly 2 layers, 1) continuous probabilities estimated from training data; 2) continuous probabilities obtained from out-of-sample (validation) data}

\item{probs}{probability levels consisting of a 'successful outcome' (e.g. presence vs. absence) from which to compute validation metrics}
}
\value{
data.frame with tabulated results
}
\description{
Compute a suite of common metrics serving to assess predictive capacity of an empirical/training model given an observed/validation sample
}
