% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmKDE-package.R
\docType{package}
\name{wmKDE-package}
\alias{wmKDE-package}
\alias{_PACKAGE}
\title{wmKDE: Weighted mean kernel density estimation}
\description{
Estimation of spatially and individually UD weighted mean bivariate population kernels
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tylerdrudolph/wmKDE}
  \item Report bugs at \url{https://github.com/tylerdrudolph/wmKDE/issues}
}

}
\author{
\strong{Maintainer}: Tyler D. Rudolph \email{tylerdrudolph@gmail.com} (\href{https://orcid.org/0000-0003-2474-7830}{ORCID})

}
