% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmKDE.R
\name{wmKDE}
\alias{wmKDE}
\title{Weighted mean kernel density estimation}
\usage{
wmKDE(
  x,
  id = NULL,
  avg = TRUE,
  spw = NULL,
  udw = NULL,
  herd.grid = NULL,
  bw.global = TRUE,
  zscale = TRUE,
  spatres = 1000,
  ktype = "iso",
  ncores = ifelse(avg, parallel::detectCores() - 1, 1),
  write2file = FALSE,
  ow = TRUE,
  writeDir = getwd(),
  fileTag = NULL
)
}
\arguments{
\item{x}{'sf' class object whose coordinates correspond to individual GPS relocations. x must be in a projected coordinate system and must contain a field name assigned to argument 'id' when avg=TRUE.}

\item{id}{optional character vector indicating column name corresponding to unique collar ID from which each relocated was transmitted. Must be provided when avg = TRUE.}

\item{avg}{logical indicating whether an averaged UD is desired. When avg=FALSE, only one kernel is generated using all relocations.}

\item{spw}{optional numeric vector of spatial relocation weights, length of which must be equal to nrow(x). Weights will be rescaled if sum(spw) != nrow(x).}

\item{udw}{optional 2-column data.frame containing individual UD weights. Must contain a field with name matching argument 'id' (one row per unique record) and a second field of weights entitled 'w'.}

\item{herd.grid}{optional specification of grid over which to estimate the UD. Default is kernel.grid(locs = st_coordinates(x), exp.range = 3, cell.size = spatres).}

\item{bw.global}{logical indicating whether bandwidth smoothing should be derived from all relocations (recommended) or made to vary according to individual sample (point pattern) distributions. Default uses plug-in method of bandwidth selection by default (alternative options not yet implemented).}

\item{zscale}{logical indicating whether individual UD probability densities should be rescaled prior to cellwise averaging (recommended). Only applicable when avg = TRUE.}

\item{spatres}{vector of length 1 specifying the desired spatial resolution of the output UD in the x & y dimensions. Asymmetrical cells not implemented.}

\item{ktype}{character vector indicating type of kernel value to return. Options are 'iso' = isopleth contours (default, higher values indicate greater probability); 'prob' = UD probabilities; 'vol' = UD volumes (sum to 1). Multiple arguments are accepted.}

\item{ncores}{integer indicating the number of parallel processes (threads) over which to execute the estimation of individual UDs. Defaults to detectCores()-1 when avg = TRUE, otherwise 1.}

\item{write2file}{logical; write results to file?}

\item{ow}{logical; overwrite existing files?}

\item{writeDir}{optional path to desired write folder location. Default is working directory.}

\item{fileTag}{optional string to append to file name when export=TRUE.}
}
\value{
list containing spatRaster object(s) of length equal to length(ktype) and one sf multipolygons object corresponding to isopleth and core area contours
}
\description{
Estimate the mean utilization distribution (UD) of a wildlife population given GPS collar telemetry relocations derived from individually monitored sample animals.
Two levels of weighting are possible:
\enumerate{
\item Individual GPS relocations may be weighted during kernel density estimation to account for variation in data quality and/or habitat transformation over time.
\item Individual UDs may be weighted during cell-wise averaging to account for individual variation in sample size and/or quality.
}
}
