% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmLDE.R
\name{wmLDK}
\alias{wmLDK}
\title{Weighted mean lattice density estimation for acoustic telemetry data}
\usage{
wmLDK(
  x = NULL,
  id = NULL,
  bath = NULL,
  avg = TRUE,
  spatres = 10,
  udw = NULL,
  zscale = T,
  ktype = c("iso", "prob"),
  ncores = ifelse(avg, parallelly::availableCores() - 1, 1),
  write2file = F,
  ow = F,
  fileTag = NULL,
  writeDir = getwd(),
  retObj = T,
  verbose = T
)
}
\arguments{
\item{x}{'sf' class object whose coordinates correspond to individual acoustic telemetry locations. x must be in a projected coordinate system and must contain a field name assigned to argument 'id' when avg = TRUE.}

\item{id}{optional character vector indicating column name corresponding to unique tag ID from which each location was detected by a receptor. Must be provided when avg = TRUE.}

\item{bath}{'sf' class POLYGON or MULTIPOLYGON object whose boundaries correspond to the water body under study (required).}

\item{avg}{logical indicating whether an averaged UD is desired. When avg = FALSE, only one kernel is generated using all detections.}

\item{spatres}{size of lattice mesh in metres (numeric of length 1)}

\item{udw}{optional 2-column data.frame containing individual UD weights. Must contain a field with name matching argument 'id' (one row per unique record) and a second field of weights entitled 'w'.}

\item{zscale}{logical indicating whether individual UD probability densities should be rescaled prior to cellwise averaging (recommended). Only applicable when avg = TRUE.}

\item{ktype}{character vector indicating type of kernel value to return. Options are 'iso' = isopleth contours (default, higher values indicate greater probability); 'prob' = UD probabilities; and 'vol' = UD volumes (sum to 1). Multiple arguments are accepted.}

\item{ncores}{integer indicating the number of parallel processes (threads) over which to execute the estimation of individual UDs. Defaults to parallelly::availableCores()-1 when avg = TRUE, otherwise 1.}

\item{write2file}{logical; write results to file?}

\item{ow}{logical; overwrite existing files?}

\item{fileTag}{optional string to append to file name when export=TRUE.}

\item{writeDir}{optional path to desired write folder location. Default is working directory.}

\item{retObj}{logical; should result be returned into R environment?}

\item{verbose}{logical; should status messages be printed?}
}
\value{
list containing 1) spatRaster object(s) of length equal to length(ktype), 2) one sf multipolygons object corresponding to isopleth, 3) core area contours and 4)
}
\description{
Weighted mean lattice density estimation for acoustic telemetry data
}
