% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmUD.R
\name{wmUD}
\alias{wmUD}
\title{Cell-wise weighted means of multiple Utilization Distributions}
\usage{
wmUD(udList, w, sproj = NULL, silent = FALSE, dres = NULL, checksum = TRUE)
}
\arguments{
\item{udList}{list of utilization distributions of equal dimensions.}

\item{w}{vector of individual UD weights for cell-wise weighted means.}

\item{sproj}{projected coordinate system of input UDs}

\item{silent}{logical; should messages be printed?}

\item{dres}{desired pixel resolution of output UD}

\item{checksum}{logical; should adjustments be made to ensure sure that all cell volumes sum to 1?}
}
\value{
weighted mean spatRaster object
}
\description{
Cell-wise weighted means of multiple Utilization Distributions
}
