% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bKDE.R
\name{bKDE}
\alias{bKDE}
\title{Batch deployment of multiple kernel density estimations based on levels of a specified factor.}
\usage{
bKDE(
  xy,
  id,
  wts = NULL,
  ncores = parallel::detectCores() - 1,
  user.grid = NULL,
  bw.global = TRUE,
  write2file = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{xy}{2-column matrix or data.frame defining projected x and y coordinates of individual GPS relocations.}

\item{id}{vector defining unique sample animals from which individual GPS relocations were derived.}

\item{wts}{vector defining weights to be attributed to individual relocations. If the sum of relocation weights does not equal the sample size, weights are adjusted accordingly and a warning is issued.}

\item{ncores}{number of threads (processes) over which to dispatch individual kernel density estimators.}

\item{user.grid}{dimensions of the matrix over which the kernel will be estimated. Mandatory.}

\item{bw.global}{logical indicating whether bandwidth smoothing should be derived from all relocations (recommended) or made to vary according to individual sample (point pattern) distributions.}

\item{write2file}{logical indicating whether results should be written to file}

\item{verbose}{logical indicating whether messages should be printed}
}
\value{
list of UD objects (list of x1, x2, fhat) of length equal to length(unique(id))
}
\description{
Batch deployment of multiple kernel density estimations based on levels of a specified factor.
}
