% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.area.R
\name{core.area}
\alias{core.area}
\title{Derive the core area isopleth from a utilization distribution using the method described in Vander Wal & Rogers (2012).}
\usage{
core.area(UD, plot = FALSE, silent = TRUE)
}
\arguments{
\item{UD}{list containing 3 elements: 'x1' (x coordinates), 'x2' (y coordinates) and 'fhat' (probability densities)}

\item{plot}{logical indicating whether or not to print result}

\item{silent}{logical indicating whether or not to print messages (default = TRUE)}
}
\description{
Derive the core area isopleth from a utilization distribution using the method described in Vander Wal & Rogers (2012).
}
