% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1score.R
\name{F1score}
\alias{F1score}
\title{F1 Score for binned continuous data (kernel utilization distributions)}
\usage{
F1score(pred, obs, id = NULL, threshVal = 70, binWidth = 100, beta = 1)
}
\arguments{
\item{pred}{spatRaster object corresponding to predicted probabilities of occurrence (wmKDE()$iso) from fitted wmKDE (training set) model (0 <= x >= 100).}

\item{obs}{spatRaster object corresponding to observed probabilities of occurrence (wmKDE()$iso) from fitted wmKDE (validation set) model (0 <= x >= 100). Alternatively, can be an sf object ('POINT' or 'MULTIPOINT') corresponding to 'true' observations.}

\item{id}{for weighted measures; character vector of length one corresponding to the field in obs (if SpatialPoints) distinguishing factor levels, in which case a confusion table will be constructed for each level of the factor and weighted according to the number of points in each level.}

\item{threshVal}{minimum probability value considered as a 'presence' (TRUE).}

\item{binWidth}{range of values per bin on a scale of 100. Default value of 100 will result in two bins: >= threshVal & > threshVal.}

\item{beta}{relative weight of recall relative to precision. Default is equal weight to both.}
}
\value{
either a list containing 1) recall, 2) precision, and 3) F1 score for the evaluated data (predicted vs observed) and input parameters, or a data.frame if obs is binary and !is.null(id).
}
\description{
F1 Score for binned continuous data (kernel utilization distributions)
}
