% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhat2confin.R
\name{fhat2confin}
\alias{fhat2confin}
\title{Convert UD density estimates into relative probabilities}
\usage{
fhat2confin(r, ...)
}
\arguments{
\item{r}{spatRaster corresponding to UD kernel density values}
}
\value{
spatRaster correponding to kernel probability isopleths
}
\description{
Convert UD density estimates into relative probabilities
}
