% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast2UD.R
\name{rast2UD}
\alias{rast2UD}
\title{Convert a spatRaster object to a list representing a UD (utilization distribution) object (list of x1, x2, fhat)}
\usage{
rast2UD(r, sproj = NULL)
}
\arguments{
\item{r}{spatRaster object}

\item{sproj}{projected CRS of input spatRaster}
}
\value{
UD object retaining cell values of original spatRaster object at original grid cell locations.
}
\description{
Convert a spatRaster object to a list representing a UD (utilization distribution) object (list of x1, x2, fhat)
}
